+++ ls_Toasts/systems/activities.lua
@@ -82,7 +82,7 @@ E:RegisterOptions("activities", {
 				else
 					Disable()
 				end
-			end
+			end,
 		},
 		dnd = {
 			order = 2,